data-science-platform/requirements.txt
bcrypt==4.0.1
fastapi==0.119.1
httpx==0.28.1
joblib==1.4.2
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.11
PyJWT==2.10.1
pydantic==2.12.3
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
scikit-learn==1.5.1
streamlit==1.50.0
uvicorn==0.38.0
